dev-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile dev-requirements.in
#
annotated-types==0.7.0
    # via pydantic
black==24.10.0
    # via -r dev-requirements.in
click==8.1.8
    # via black
flake8==7.1.1
    # via -r dev-requirements.in
gitdb==4.0.12
    # via
    #   -r dev-requirements.in
    #   gitpython
gitpython==3.1.44
    # via -r dev-requirements.in
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via -r dev-requirements.in
jinja2==3.1.5
    # via -r /workspaces/generator-terraform/requirements.in
mako==1.3.8
    # via pdoc3
markdown==3.7
    # via pdoc3
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
packaging==24.2
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via black
pdoc3==0.11.5
    # via -r dev-requirements.in
platformdirs==4.3.6
    # via black
pluggy==1.5.0
    # via pytest
pycodestyle==2.12.1
    # via flake8
pydantic==2.10.5
    # via -r /workspaces/generator-terraform/requirements.in
pydantic-core==2.27.2
    # via pydantic
pyflakes==3.2.0
    # via flake8
pytest==8.3.4
    # via -r dev-requirements.in
pyyaml==6.0.2
    # via -r /workspaces/generator-terraform/requirements.in
smmap==5.0.2
    # via gitdb
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
